视频列表显示视频时长

+++ bilimiao-comm/src/main/java/com/a10miaomiao/bilimiao/comm/entity/region/RegionTypeDetailsInfo.kt
@@ -18,5 +18,5 @@ data class RegionTypeDetailsInfo(
     var play: String,
     var title: String,
     var video_review: String,
-    var duration: Int,
+    var duration: Long,
 )
